Rename and split in two versions
